испраил парсер, раньше было мега тупо сейчас не мега тупо: collectProductInfo принимает String html и возвращает инфу о товаре, ExtractHtml принимает ссылку, возвращает html. Тоесть получить инфу о товаре по ссылке теперь так: ProductInfoCollector.collectProductInfo(FetchHtml.ExtractHtml("ссылка")). Пока возращает только цену без озон карты, и айди. Делаю остальное.
